refactor: new pkg wayidsflag
